ENH,MAINT: Allow cupy-like protocols in setitem and consolidate
This allows `cupy_arr[...] = cupy_like` where cupy-like supports the
CUDA array interface or `__cupy_get_ndarray__`.

Since there were a few (not many!) places that have a similar logic
I, opionatedly, consolidated them into a `cdef inline` function.
Similarly moved the test helper classes to not have to repeat them.

The biggest part of the diff is just moving those helpers!

I am very sure that `cdef inline` must be in the `.pyd` file to
actually be inlined (which means many are currenlty not).
