+++ tests/cupy_tests/core_tests/test_ndarray_indexing.py
@@ -8,6 +8,8 @@ import pytest
 
 import cupy
 from cupy import testing
+from cupy.testing._protocol_helpers import (
+    DummyObjectWithCuPyGetNDArray, DummyObjectWithCudaArrayInterface)
 from cupy.exceptions import ComplexWarning
 
 
@@ -233,3 +235,14 @@ class TestSetItemCompatBroadcast:
         a = xp.zeros((2, 3, 4), dtype)
         a[0, 1, 2] = testing.shaped_arange((), xp, dtype)
         return a
+
+
+@pytest.mark.parametrize('cupy_like', [
+    DummyObjectWithCuPyGetNDArray,
+    DummyObjectWithCudaArrayInterface,
+])
+def test_setitem_with_cupy_like(cupy_like):
+    # Test that normal assignment supports interfaces
+    a = cupy.zeros(10)
+    a[...] = cupy_like(cupy.arange(10))
+    testing.assert_array_equal(a, cupy.arange(10))
